adding confirmation for critical operations like delete/clear/schema removal

+++ src/here-xyz.ts
@@ -60,6 +60,14 @@ const questions = [
     }
 ];
 
+const questionConfirm = [
+    {
+        type: 'input',
+        name: 'confirmed',
+        message: 'Enter (Y)es to continue or (N)o to cancel'
+    }
+];
+
 const questionAnalyze = [
     {
         type: "checkbox",
@@ -847,6 +855,16 @@ program
     .description("delete the xyzspace with the given id")
     .action(async geospaceId => {
         //console.log("geospaceId:"+"/geospace/"+geospaceId);
+
+    console.log("Are you sure you want to delete the given space ?");
+    const answer = await inquirer.prompt<{ confirmed?: string }>(questionConfirm);
+
+    const termsResp = answer.confirmed ? answer.confirmed.toLowerCase() : 'no';
+    if (termsResp !== "y" && termsResp !== "yes") {
+        console.log("CANCELLED !");
+        process.exit(1);
+    }
+
         deleteSpace(geospaceId)
             .catch((error) => {
                 if (error.statusCode == 404) {
@@ -859,6 +877,8 @@ program
     });
 
 async function deleteSpace(geospaceId: string) {
+    
+    
     const { response, body } = await execute(
         "/hub/spaces/" + geospaceId + "?clientId=cli",
         "DELETE",
@@ -928,14 +948,23 @@ program
     .description("clear data from xyz space")
     .option("-t, --tags [tags]", "tags for the xyz space")
     .option("-i, --ids [ids]", "ids for the xyz space")
-    .action((id, options) => {
+    .action( async (id, options) => {        
+        console.log("Are you sure you want to clear data of the given space ?");
+        const answer = await inquirer.prompt<{ confirmed?: string }>(questionConfirm);
+    
+        const termsResp = answer.confirmed ? answer.confirmed.toLowerCase() : 'no';
+        if (termsResp !== "y" && termsResp !== "yes") {
+            console.log("CANCELLED !");
+            process.exit(1);
+        }
+
         clearSpace(id, options).catch((error) => {
             if (error.statusCode == 404) {
                 console.log("Operation FAILED: space does not exist");
             } else {
                 handleError(error);
             }
-        })
+        })    
     });
 
 async function clearSpace(id: string, options: any) {
@@ -1794,10 +1823,18 @@ async function configXyzSpace(id:string, options:any) {
     }
 
     if(options.schema) {
-        
         if(options.schema == true) {
+            console.log("Are you sure you want to remove the schema definition of the given space ?");
+            const answer = await inquirer.prompt<{ confirmed?: string }>(questionConfirm);
+
+            const termsResp = answer.confirmed ? answer.confirmed.toLowerCase() : 'no';
+            if (termsResp !== "y" && termsResp !== "yes") {
+                console.log("CANCELLED !");
+                process.exit(1);
+            }
             console.log("Removing schema def for the space.")
             patchRequest['processors'] = [];
+
         } else {
             let schemaDef:string = "";
             if(options.schema.indexOf("http") == 0) {
